Massive revamp of the systemd interface for NixOS
We finally managed to get standard systemd units (e.g. services and timers) to work on NixOS without having to go through nixos-rebuild (although we also added an interface for that). The challenge with NixOS is that it doesn't let us write to /etc/systemd/system which is necessary for 'systemctl enable'. Therefore we can't enable and run our own systemd system units.

The solution was to use a systemd 'generator' that is registered with NixOS. It's a script that runs before systemd looks at all the unit files during boot and therefore we can use it to generate/copy our own unit files. One challenge was that it doesn't 'systemmctl enable' the units so we had to implement that ourselves (create symlink paths for .wants etc).